--- a/util.go
+++ b/util.go
@@ -292,3 +292,24 @@ func minOfUint16(a, b uint16) uint16 {
 	}
 	return b
 }
+
+func maxInt(a, b int) int {
+	if a > b {
+		return a
+	}
+	return b
+}
+
+func maxUint16(a, b uint16) uint16 {
+	if a > b {
+		return a
+	}
+	return b
+}
+
+func minUint16(a, b uint16) uint16 {
+	if a < b {
+		return a
+	}
+	return b
+}