This PR introduces a faster _ParOr_. The algorithm is described informally in #140.

Benchmark data shows significant improvement compared to previous algorithm.

Benchmarks run on 15' 2017 MacBook Pro Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz
```
benchmark                                           old ns/op     new ns/op     delta
BenchmarkRealDataParOr/census-income_srt-8          270110        198445        -26.53%
BenchmarkRealDataParOr/census-income-8              383111        302936        -20.93%
BenchmarkRealDataParOr/census1881_srt-8             679945        404061        -40.57%
BenchmarkRealDataParOr/census1881-8                 701318        606377        -13.54%
BenchmarkRealDataParOr/dimension_003-8              7206730       2208563       -69.35%
BenchmarkRealDataParOr/dimension_008-8              3016907       959910        -68.18%
BenchmarkRealDataParOr/dimension_033-8              469810        201631        -57.08%
BenchmarkRealDataParOr/uscensus2000-8               1882977       1081049       -42.59%
BenchmarkRealDataParOr/weather_sept_85_srt-8        382546        158591        -58.54%
BenchmarkRealDataParOr/weather_sept_85-8            1255767       1240276       -1.23%
BenchmarkRealDataParOr/wikileaks-noquotes_srt-8     286285        123959        -56.70%
BenchmarkRealDataParOr/wikileaks-noquotes-8         328856        166740        -49.30%
```

The previous algorithm is exposed in the public API as ParHeapOr somehow resembling the convention in `fastaggregation.go`.

Resolves #140