@maciej 2018-04-03
Whoops! I haven't done proper testing before pushing this! Sorry about that! It should be a trivial edge case to fix, let me do just that!