@maciej 2018-04-03
Bizarrely those tests fail on Travis and pass on my local machine.
```
$ go test
--- FAIL: TestParAggregations (0.05s)
    --- FAIL: TestParAggregations/advanced_case (0.04s)
    	aggregation_test.go:218: Expected bitmap of cardinality 857144, got 758346
    --- FAIL: TestParAggregations/advanced_case_with_runs (0.02s)
    	aggregation_test.go:260: Expected bitmap of cardinality 635287, got 536489
```

I've tested it against Go 1.10.1, 1.9.5 and 1.9.4. In each case the cardinality is the same.