@maciej 2018-04-03
OK, I'm able to reproduce it on `linux/arm` (Raspberry Pi)
```
◦  go test -run TestParAgg                                                                                         220:47:19
--- FAIL: TestParAggregations (0.37s)
    --- FAIL: TestParAggregations/advanced_case (0.20s)
    	aggregation_test.go:218: Expected bitmap of cardinality 857144, got 477478
    --- FAIL: TestParAggregations/advanced_case_with_runs (0.16s)
    	aggregation_test.go:260: Expected bitmap of cardinality 635287, got 255621
FAIL
exit status 1
FAIL	github.com/RoaringBitmap/roaring	0.383s
```