@lemire 2018-04-03
If you’d like, I can do a code review now or do you prefer I wait?