@lemire 2018-04-03
cc @richardstartin