@maciej 2018-04-03
@lemire please wait. I messed up chunk size calculation. Due to some tuning constants the tests passed on a machine with `NumCPU = 8` (4 core Mac with hyper threading), but failed on eg. `parallelism = 4`. I'll improve the test coverage to check against that and get back to you!

Unfortunately that only means some of the benchmark results were wrong. I won't post the updated results just yet, I want to be 100% sure they are correct.