@richardstartin 2018-04-03
One thing to pay attention to might be `uscenus2000` - that's over 1ms, and is under 100us in a [Java PR](https://github.com/RoaringBitmap/RoaringBitmap/pull/234), on my slightly lower spec machine. I would be interested to see these two approaches implemented in the same language, in each language, to better understand the tradeoff between up front work and balance.