@maciej 2018-04-04
@richardstartin thank you for sharing the dataset characteristics! These are very valuable!
Yes, I think it'd make sense to have all _ParOr_ implementations compared against one another at some point!
For now I would not rely on the numbers I posted above. They are wrong.