@maciej 2018-04-04
I'm closing this PR. I'll do #168 and #169 and issue another PR once I'm confident that it's correct.